move: Room 클래스를 하위 패키지로 분류
새로운 Room 테이블을 선언함에 따라 기존에 존재하던 RoomVocabulary를 별도의 패키지로 분류하였다. 앞으로 테이블을 하나 선언할 때마다 패키지를 새로 만들자.
