+++ app/src/androidTest/java/hsk/practice/myvoca/room/ObjectConvertTest.kt
@@ -5,6 +5,7 @@ import com.google.gson.Gson
 import hsk.practice.myvoca.data.MeaningImpl
 import hsk.practice.myvoca.data.VocabularyImpl
 import hsk.practice.myvoca.data.WordClassImpl
+import hsk.practice.myvoca.room.vocabulary.*
 import junit.framework.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
