+++ app/src/androidTest/java/hsk/practice/myvoca/room/VocaDaoTest.kt
@@ -8,6 +8,10 @@ import hsk.practice.myvoca.data.MeaningImpl
 import hsk.practice.myvoca.data.VocabularyImpl
 import hsk.practice.myvoca.data.WordClassImpl
 import hsk.practice.myvoca.data.fakeData
+import hsk.practice.myvoca.room.vocabulary.RoomVocaDatabase
+import hsk.practice.myvoca.room.vocabulary.VocaDao
+import hsk.practice.myvoca.room.vocabulary.toRoomVocabulary
+import hsk.practice.myvoca.room.vocabulary.toVocabularyImpl
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.runBlocking
 import org.junit.After
