+++ app/src/main/java/hsk/practice/myvoca/module/RoomVocaDatabaseModule.kt
@@ -7,9 +7,9 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import hsk.practice.myvoca.room.RoomMigrations
-import hsk.practice.myvoca.room.RoomVocaDatabase
-import hsk.practice.myvoca.room.VocaDao
+import hsk.practice.myvoca.room.vocabulary.RoomMigrations
+import hsk.practice.myvoca.room.vocabulary.RoomVocaDatabase
+import hsk.practice.myvoca.room.vocabulary.VocaDao
 import javax.inject.Singleton
 
 @Module
