+++ app/src/main/java/hsk/practice/myvoca/room/vocabulary/VocaDao.kt
@@ -1,4 +1,4 @@
-package hsk.practice.myvoca.room
+package hsk.practice.myvoca.room.vocabulary
 
 import androidx.room.*
 import kotlinx.coroutines.flow.Flow
